demo/requirements.txt
accelerate==1.10.1
av==15.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.13.1
fsspec==2024.6.1
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==7.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
peft==0.17.1
pillow==11.0.0
psutil==7.0.0
pycocotools==2.0.10
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
sympy==1.13.3
tokenizers==0.22.0
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.56.0
triton==3.4.0
typing_extensions==4.12.2
urllib3==2.5.0
zipp==3.19.2
